+++ source/dynamical_systems/test/tests/test_ring.cpp
@@ -1,21 +1,28 @@
 #include <gtest/gtest.h>
-#include "dynamical_systems/Ring.hpp"
+
+#include "dynamical_systems/DynamicalSystemFactory.hpp"
 #include "dynamical_systems/exceptions/EmptyBaseFrameException.hpp"
 #include "dynamical_systems/exceptions/EmptyAttractorException.hpp"
 
+#include "state_representation/parameters/Parameter.hpp"
 #include "state_representation/exceptions/IncompatibleReferenceFramesException.hpp"
 #include "state_representation/exceptions/EmptyStateException.hpp"
 
+using namespace dynamical_systems;
 using namespace state_representation;
 using namespace std::literals::chrono_literals;
 
 class RingDSTest : public ::testing::Test {
 protected:
   RingDSTest() {
+    ds = DynamicalSystemFactory<CartesianState>::create_dynamical_system(
+        DynamicalSystemFactory<CartesianState>::DYNAMICAL_SYSTEM::RING
+    );
     center = CartesianPose::Identity("A");
     current_pose = CartesianPose("B", radius * Eigen::Vector3d::Random());
   }
 
+  std::shared_ptr<IDynamicalSystem<CartesianState>> ds;
   CartesianPose current_pose;
   CartesianPose center;
   std::vector<double> vlim = {10, 10, 0.001, 0.001};
@@ -29,130 +36,137 @@ protected:
 };
 
 TEST_F(RingDSTest, EmptyConstructor) {
-  // construct empty cartesian state DS
   CartesianPose center = CartesianPose::Identity("CAttractor", "A");
-  dynamical_systems::Ring ds;
   // base frame and attractor should be empty
-  EXPECT_TRUE(ds.get_center().is_empty());
-  EXPECT_TRUE(ds.get_base_frame().is_empty());
-  ds.set_center(center);
-  EXPECT_FALSE(ds.get_center().is_empty());
-  EXPECT_FALSE(ds.get_base_frame().is_empty());
+  EXPECT_TRUE(ds->get_parameter_value<CartesianPose>("center").is_empty());
+  ds->set_parameter(make_shared_parameter("center", center));
+  EXPECT_FALSE(ds->get_parameter_value<CartesianPose>("center").is_empty());
+  EXPECT_FALSE(ds->get_base_frame().is_empty());
   // when attractor was set without a base frame, expect base frame to be identity with name / reference_frame of attractor
-  EXPECT_EQ(ds.get_base_frame().get_name(), center.get_reference_frame());
-  EXPECT_EQ(ds.get_base_frame().get_reference_frame(), center.get_reference_frame());
-  EXPECT_EQ(ds.get_base_frame().get_transformation_matrix(), Eigen::Matrix4d::Identity());
+  EXPECT_EQ(ds->get_base_frame().get_name(), center.get_reference_frame());
+  EXPECT_EQ(ds->get_base_frame().get_reference_frame(), center.get_reference_frame());
+  EXPECT_EQ(ds->get_base_frame().get_transformation_matrix(), Eigen::Matrix4d::Identity());
+}
 
-  ds = dynamical_systems::Ring();
+TEST_F(RingDSTest, EvaluateComptability) {
+  CartesianPose center = CartesianPose::Identity("CAttractor", "A");
   CartesianState state1 = CartesianState::Identity("B", "A");
   CartesianState state2("D", "C");
   CartesianState state3("C", "A");
   CartesianState state4("C", "B");
   // if no base frame is set, an exception is thrown
-  EXPECT_THROW(ds.evaluate(state1), dynamical_systems::exceptions::EmptyBaseFrameException);
-  ds.set_base_frame(state1);
+  EXPECT_THROW(ds->evaluate(state1), dynamical_systems::exceptions::EmptyBaseFrameException);
+  ds->set_base_frame(state1);
   // if cartesian state is incompatible, an exception is thrown
-  EXPECT_THROW(ds.evaluate(state2), state_representation::exceptions::IncompatibleReferenceFramesException);
+  EXPECT_THROW(ds->evaluate(state2), state_representation::exceptions::IncompatibleReferenceFramesException);
   // if cartesian state needs to be transformed to other frame first and is empty, an exception is thrown
-  EXPECT_THROW(ds.evaluate(state3), state_representation::exceptions::EmptyStateException);
+  EXPECT_THROW(ds->evaluate(state3), state_representation::exceptions::EmptyStateException);
   // if no attractor is set, an exception is thrown
-  EXPECT_THROW(ds.evaluate(state4), dynamical_systems::exceptions::EmptyAttractorException);
+  EXPECT_THROW(ds->evaluate(state4), dynamical_systems::exceptions::EmptyAttractorException);
 
-  ds.set_center(center);
-  EXPECT_TRUE(ds.is_compatible(state1));
-  EXPECT_FALSE(ds.is_compatible(state2));
-  EXPECT_TRUE(ds.is_compatible(state3));
-  EXPECT_TRUE(ds.is_compatible(state4));
+  ds->set_parameter(make_shared_parameter("center", center));
+  EXPECT_TRUE(ds->is_compatible(state1));
+  EXPECT_FALSE(ds->is_compatible(state2));
+  EXPECT_TRUE(ds->is_compatible(state3));
+  EXPECT_TRUE(ds->is_compatible(state4));
 }
 
 TEST_F(RingDSTest, PointsOnRadius) {
-  dynamical_systems::Ring ring(center, radius, width, speed);
+  ds->set_parameter(make_shared_parameter<CartesianPose>("center", center));
+  ds->set_parameter(make_shared_parameter<double>("radius", radius));
+  ds->set_parameter(make_shared_parameter<double>("width", width));
+  ds->set_parameter(make_shared_parameter<double>("speed", speed));
   CartesianTwist twist;
 
   // zero output at center
   current_pose = center;
-  twist = ring.evaluate(current_pose);
+  twist = ds->evaluate(current_pose);
   EXPECT_NEAR(twist.data().norm(), 0, tol);
 
   current_pose.set_position(radius, 0, 0);
-  twist = ring.evaluate(current_pose);
+  twist = ds->evaluate(current_pose);
   EXPECT_NEAR(twist.get_linear_velocity().x(), 0, tol);
   EXPECT_NEAR(twist.get_linear_velocity().y(), speed, tol);
 
   current_pose.set_position(0, radius, 0);
-  twist = ring.evaluate(current_pose);
+  twist = ds->evaluate(current_pose);
   EXPECT_NEAR(twist.get_linear_velocity().x(), -speed, tol);
   EXPECT_NEAR(twist.get_linear_velocity().y(), 0, tol);
 
   current_pose.set_position(-radius, 0, 0);
-  twist = ring.evaluate(current_pose);
+  twist = ds->evaluate(current_pose);
   EXPECT_NEAR(twist.get_linear_velocity().x(), 0, tol);
   EXPECT_NEAR(twist.get_linear_velocity().y(), -speed, tol);
 
   current_pose.set_position(0, -radius, 0);
-  twist = ring.evaluate(current_pose);
+  twist = ds->evaluate(current_pose);
   EXPECT_NEAR(twist.get_linear_velocity().x(), speed, tol);
   EXPECT_NEAR(twist.get_linear_velocity().y(), 0, tol);
 
   current_pose.set_position(radius, 0, 1);
-  twist = ring.evaluate(current_pose);
+  twist = ds->evaluate(current_pose);
   EXPECT_NEAR(twist.get_linear_velocity().z(), -1, tol);
 }
 
 TEST_F(RingDSTest, PointsNearRadius) {
-  dynamical_systems::Ring ring(center, radius, width, speed, field_strength);
+  ds->set_parameter(make_shared_parameter<CartesianPose>("center", center));
+  ds->set_parameter(make_shared_parameter<double>("radius", radius));
+  ds->set_parameter(make_shared_parameter<double>("width", width));
+  ds->set_parameter(make_shared_parameter<double>("speed", speed));
+  ds->set_parameter(make_shared_parameter<double>("field_strength", field_strength));
   CartesianTwist twist;
 
   current_pose.set_position(radius + width, 0, 0);
-  twist = ring.evaluate(current_pose);
+  twist = ds->evaluate(current_pose);
   EXPECT_NEAR(twist.get_linear_velocity().x(), -speed * field_strength, tol);
   EXPECT_NEAR(twist.get_linear_velocity().y(), 0, tol);
 
   current_pose.set_position(radius - width, 0, 0);
-  twist = ring.evaluate(current_pose);
+  twist = ds->evaluate(current_pose);
   EXPECT_NEAR(twist.get_linear_velocity().x(), speed * field_strength, tol);
   EXPECT_NEAR(twist.get_linear_velocity().y(), 0, tol);
 
   current_pose.set_position(0, radius + width, 0);
-  twist = ring.evaluate(current_pose);
+  twist = ds->evaluate(current_pose);
   EXPECT_NEAR(twist.get_linear_velocity().x(), 0, tol);
   EXPECT_NEAR(twist.get_linear_velocity().y(), -speed * field_strength, tol);
 
   current_pose.set_position(0, radius - width, 0);
-  twist = ring.evaluate(current_pose);
+  twist = ds->evaluate(current_pose);
   EXPECT_NEAR(twist.get_linear_velocity().x(), 0, tol);
   EXPECT_NEAR(twist.get_linear_velocity().y(), speed * field_strength, tol);
 
   current_pose.set_position(radius + (width * 2), 0, 0);
-  twist = ring.evaluate(current_pose);
+  twist = ds->evaluate(current_pose);
   EXPECT_NEAR(twist.get_linear_velocity().x(), -speed * field_strength, tol);
   EXPECT_NEAR(twist.get_linear_velocity().y(), 0, tol);
 }
 
 TEST_F(RingDSTest, BehaviourNearBoundaryAtHighSpeeds) {
-  radius = 1;
-  width = 0.1;
-  speed = 10;
-  field_strength = 2;
-  dynamical_systems::Ring ring(center, radius, width, speed, field_strength);
+  ds->set_parameter(make_shared_parameter<CartesianPose>("center", center));
+  ds->set_parameter(make_shared_parameter<double>("radius", 1.0));
+  ds->set_parameter(make_shared_parameter<double>("width", 0.1));
+  ds->set_parameter(make_shared_parameter<double>("speed", 10.0));
+  ds->set_parameter(make_shared_parameter<double>("field_strength", 2.0));
   CartesianTwist twist;
 
   current_pose.set_position(0, radius + 1.001 * width, 0);
-  twist = ring.evaluate(current_pose);
+  twist = ds->evaluate(current_pose);
   double angular_speed = twist.get_angular_velocity().norm();
   current_pose.set_position(0, radius + 1.000 * width, 0);
-  twist = ring.evaluate(current_pose);
+  twist = ds->evaluate(current_pose);
   EXPECT_NEAR(twist.get_angular_velocity().norm(), angular_speed, 0.1);
   current_pose.set_position(0, radius + 0.999 * width, 0);
-  twist = ring.evaluate(current_pose);
+  twist = ds->evaluate(current_pose);
   EXPECT_NEAR(twist.get_angular_velocity().norm(), angular_speed, 0.1);
 }
 
 TEST_F(RingDSTest, ConvergenceOnRadius) {
-  dynamical_systems::Ring ring(center, radius);
+  ds->set_parameter(make_shared_parameter<CartesianPose>("center", center));
+  ds->set_parameter(make_shared_parameter<double>("radius", radius));
 
   for (unsigned int i = 0; i < nb_steps; ++i) {
-    CartesianTwist twist = ring.evaluate(current_pose);
+    CartesianTwist twist = ds->evaluate(current_pose);
     twist.clamp(vlim[0], vlim[1], vlim[2], vlim[3]);
     current_pose += dt * twist;
   }
@@ -163,10 +177,11 @@ TEST_F(RingDSTest, ConvergenceOnRadius) {
 TEST_F(RingDSTest, ConvergenceOnRadiusRandomCenter) {
   center.set_position(Eigen::Vector3d::Random());
   center.set_orientation(Eigen::Quaterniond::UnitRandom());
-  dynamical_systems::Ring ring(center, radius);
+  ds->set_parameter(make_shared_parameter<CartesianPose>("center", center));
+  ds->set_parameter(make_shared_parameter<double>("radius", radius));
 
   for (unsigned int i = 0; i < nb_steps; ++i) {
-    CartesianTwist twist = ring.evaluate(current_pose);
+    CartesianTwist twist = ds->evaluate(current_pose);
     twist.clamp(vlim[0], vlim[1], vlim[2], vlim[3]);
     current_pose += dt * twist;
   }
@@ -175,12 +190,12 @@ TEST_F(RingDSTest, ConvergenceOnRadiusRandomCenter) {
 }
 
 TEST_F(RingDSTest, ZeroNormalGain) {
-  dynamical_systems::Ring ring(center);
-  ring.set_normal_gain(0);
+  ds->set_parameter(make_shared_parameter<CartesianPose>("center", center));
+  ds->set_parameter(make_shared_parameter<double>("normal_gain", 0.0));
 
   double startingHeight = current_pose.get_position().z();
   for (unsigned int i = 0; i < nb_steps; ++i) {
-    CartesianTwist twist = ring.evaluate(current_pose);
+    CartesianTwist twist = ds->evaluate(current_pose);
     twist.clamp(vlim[0], vlim[1], vlim[2], vlim[3]);
     current_pose += dt * twist;
   }
@@ -188,37 +203,43 @@ TEST_F(RingDSTest, ZeroNormalGain) {
 }
 
 TEST_F(RingDSTest, OrientationAroundCircle) {
-  dynamical_systems::Ring ring(center, radius, width, 0);
+  ds->set_parameter(make_shared_parameter<CartesianPose>("center", center));
+  ds->set_parameter(make_shared_parameter<double>("radius", radius));
+  ds->set_parameter(make_shared_parameter<double>("width", width));
+  ds->set_parameter(make_shared_parameter<double>("speed", 0.0));
   CartesianTwist twist;
 
   // at the position {radius, 0, 0}, the orientation attractor is by default null,
   // so there should be zero angular velocity if the current pose orientation is also null
   current_pose.set_position(radius, 0, 0);
   current_pose.set_orientation(Eigen::Quaterniond::Identity());
-  twist = ring.evaluate(current_pose);
+  twist = ds->evaluate(current_pose);
   EXPECT_NEAR(twist.get_angular_velocity().norm(), 0, tol);
 
   // at the position {0, radius, 0}, the default rotation around Z is pi/2
   current_pose.set_position(0, radius, 0);
   current_pose.set_orientation(Eigen::Quaterniond(1, 0, 0, 1));
-  twist = ring.evaluate(current_pose);
+  twist = ds->evaluate(current_pose);
   EXPECT_NEAR(twist.get_angular_velocity().norm(), 0, tol);
 
   // at the position {0, radius, 0}, the default rotation around Z is pi
   current_pose.set_position(-radius, 0, 0);
   current_pose.set_orientation(Eigen::Quaterniond(0, 0, 0, 1));
-  twist = ring.evaluate(current_pose);
+  twist = ds->evaluate(current_pose);
   EXPECT_NEAR(twist.get_angular_velocity().norm(), 0, tol);
 
   // at the position {0, radius, 0}, the default rotation around Z is 3 * pi/2
   current_pose.set_position(0, -radius, 0);
   current_pose.set_orientation(Eigen::Quaterniond(1, 0, 0, -1));
-  twist = ring.evaluate(current_pose);
+  twist = ds->evaluate(current_pose);
   EXPECT_NEAR(twist.get_angular_velocity().norm(), 0, tol);
 }
 
 TEST_F(RingDSTest, OrientationRestitutionAtZeroAngle) {
-  dynamical_systems::Ring ring(center, radius, width, 0);
+  ds->set_parameter(make_shared_parameter<CartesianPose>("center", center));
+  ds->set_parameter(make_shared_parameter<double>("radius", radius));
+  ds->set_parameter(make_shared_parameter<double>("width", width));
+  ds->set_parameter(make_shared_parameter<double>("speed", 0.0));
   CartesianTwist twist;
 
   // at the position {radius, 0, 0}, the orientation attractor is by default null (angle around circle is 0)
@@ -226,28 +247,31 @@ TEST_F(RingDSTest, OrientationRestitutionAtZeroAngle) {
 
   // rotate the current pose around local X
   current_pose.set_orientation(Eigen::Quaterniond(1, 1, 0, 0));
-  twist = ring.evaluate(current_pose);
-  EXPECT_NEAR(twist.get_angular_velocity().x(), -M_PI_2 * ring.get_angular_gain(), tol);
+  twist = ds->evaluate(current_pose);
+  EXPECT_NEAR(twist.get_angular_velocity().x(), -M_PI_2 * ds->get_parameter_value<double>("angular_gain"), tol);
   EXPECT_NEAR(twist.get_angular_velocity().y(), 0, tol);
   EXPECT_NEAR(twist.get_angular_velocity().z(), 0, tol);
 
   // rotate the current pose around local Y
   current_pose.set_orientation(Eigen::Quaterniond(1, 0, 1, 0));
-  twist = ring.evaluate(current_pose);
+  twist = ds->evaluate(current_pose);
   EXPECT_NEAR(twist.get_angular_velocity().x(), 0, tol);
-  EXPECT_NEAR(twist.get_angular_velocity().y(), -M_PI_2 * ring.get_angular_gain(), tol);
+  EXPECT_NEAR(twist.get_angular_velocity().y(), -M_PI_2 * ds->get_parameter_value<double>("angular_gain"), tol);
   EXPECT_NEAR(twist.get_angular_velocity().z(), 0, tol);
 
   // rotate the current pose around local Z
   current_pose.set_orientation(Eigen::Quaterniond(1, 0, 0, 1));
-  twist = ring.evaluate(current_pose);
+  twist = ds->evaluate(current_pose);
   EXPECT_NEAR(twist.get_angular_velocity().x(), 0, tol);
   EXPECT_NEAR(twist.get_angular_velocity().y(), 0, tol);
-  EXPECT_NEAR(twist.get_angular_velocity().z(), -M_PI_2 * ring.get_angular_gain(), tol);
+  EXPECT_NEAR(twist.get_angular_velocity().z(), -M_PI_2 * ds->get_parameter_value<double>("angular_gain"), tol);
 }
 
 TEST_F(RingDSTest, OrientationRotationOffset) {
-  dynamical_systems::Ring ring(center, radius, width, 0);
+  ds->set_parameter(make_shared_parameter<CartesianPose>("center", center));
+  ds->set_parameter(make_shared_parameter<double>("radius", radius));
+  ds->set_parameter(make_shared_parameter<double>("width", width));
+  ds->set_parameter(make_shared_parameter<double>("speed", 0.0));
   CartesianTwist twist;
 
   current_pose.set_position(radius, 0, 0);
@@ -257,121 +281,129 @@ TEST_F(RingDSTest, OrientationRotationOffset) {
   // if the rotation offset is the same as the current orientation, the angular velocity at
   // position {radius, 0, 0} is always zero
   current_pose.set_orientation(rotation);
-  ring.set_rotation_offset(rotation);
-  twist = ring.evaluate(current_pose);
+  ds->set_parameter(make_shared_parameter("rotation_offset", current_pose));
+  twist = ds->evaluate(current_pose);
   EXPECT_NEAR(twist.get_angular_velocity().norm(), 0, tol);
 
   // now if the current pose has some orientation relative to the rotation offset,
   // it will yield the expected angular velocity of only that difference
-  current_pose.set_orientation(Eigen::Quaterniond(1, 1, 0, 0).normalized() * ring.get_rotation_offset());
-  twist = ring.evaluate(current_pose);
-  EXPECT_NEAR(twist.get_angular_velocity().x(), -M_PI_2 * ring.get_angular_gain(), tol);
+  current_pose.set_orientation(
+      Eigen::Quaterniond(1, 1, 0, 0).normalized()
+          * ds->get_parameter_value<CartesianPose>("rotation_offset").get_orientation());
+  twist = ds->evaluate(current_pose);
+  EXPECT_NEAR(twist.get_angular_velocity().x(), -M_PI_2 * ds->get_parameter_value<double>("angular_gain"), tol);
   EXPECT_NEAR(twist.get_angular_velocity().y(), 0, tol);
   EXPECT_NEAR(twist.get_angular_velocity().z(), 0, tol);
 
   // rotate the center plane in the base frame, and set the current position to have the same relative
   // offset that gives a zero command (no rotation offset)
   center.set_orientation(Eigen::Quaterniond::UnitRandom());
-  ring.set_center(center);
+  ds->set_parameter(make_shared_parameter("center", center));
   current_pose = CartesianPose("B", Eigen::Vector3d(radius, 0, 0), "A");
   current_pose = center * current_pose;
-  ring.set_rotation_offset(Eigen::Quaterniond::Identity());
-  twist = ring.evaluate(current_pose);
+  ds->set_parameter(make_shared_parameter("rotation_offset", CartesianPose::Identity("offset")));
+  twist = ds->evaluate(current_pose);
   EXPECT_NEAR(twist.get_angular_velocity().norm(), 0, tol);
 
   // for any rotation offset in a rotated center plane, the output will
   // still be zero if the current position and orientation matches the rotation offset
   rotation = Eigen::Quaterniond::UnitRandom();
-  ring.set_rotation_offset(rotation);
-  current_pose = CartesianPose("B", Eigen::Vector3d(radius, 0, 0), ring.get_rotation_offset(), "A");
+  ds->set_parameter(make_shared_parameter("rotation_offset", CartesianPose("offset", rotation)));
+  current_pose = CartesianPose(
+      "B", Eigen::Vector3d(radius, 0, 0), ds->get_parameter_value<CartesianPose>("rotation_offset").get_orientation(),
+      "A"
+  );
   current_pose = center * current_pose;
-  twist = ring.evaluate(current_pose);
+  twist = ds->evaluate(current_pose);
   EXPECT_NEAR(twist.get_angular_velocity().norm(), 0, tol);
 
   // any additional orientation in the ring frame on top of the rotation offset
   // should give the same local command, regardless of the center frame
-  current_pose = CartesianPose("B",
-                               Eigen::Vector3d(radius, 0, 0),
-                               Eigen::Quaterniond(1, 1, 0, 0).normalized() * ring.get_rotation_offset(),
-                               "A");
+  current_pose = CartesianPose(
+      "B", Eigen::Vector3d(radius, 0, 0), Eigen::Quaterniond(1, 1, 0, 0).normalized()
+          * ds->get_parameter_value<CartesianPose>("rotation_offset").get_orientation(), "A"
+  );
   current_pose = center * current_pose;
-  twist = ring.evaluate(current_pose);
+  twist = ds->evaluate(current_pose);
 
   // check the twist in the local frame
   twist = CartesianState(center).inverse() * twist;
-  EXPECT_NEAR(twist.get_angular_velocity().x(), -M_PI_2 * ring.get_angular_gain(), tol);
+  EXPECT_NEAR(twist.get_angular_velocity().x(), -M_PI_2 * ds->get_parameter_value<double>("angular_gain"), tol);
   EXPECT_NEAR(twist.get_angular_velocity().y(), 0, tol);
   EXPECT_NEAR(twist.get_angular_velocity().z(), 0, tol);
 }
 
 TEST_F(RingDSTest, BaseFrameBehaviours) {
   auto AinB = CartesianState::Random("A", "B");
-  dynamical_systems::Ring ring(AinB);
+  ds->set_parameter(make_shared_parameter<CartesianPose>("center", AinB));
 
   // setting the center through the constructor should also set the base frame (as Identity frame)
-  EXPECT_STREQ(ring.get_center().get_name().c_str(), "A");
-  EXPECT_STREQ(ring.get_center().get_reference_frame().c_str(), "B");
-  EXPECT_STREQ(ring.get_base_frame().get_name().c_str(), "B");
-  EXPECT_STREQ(ring.get_base_frame().get_reference_frame().c_str(), "B");
-  EXPECT_NEAR(ring.get_base_frame().get_pose().norm(), 1, tol);
+  EXPECT_STREQ(ds->get_parameter_value<CartesianPose>("center").get_name().c_str(), "A");
+  EXPECT_STREQ(ds->get_parameter_value<CartesianPose>("center").get_reference_frame().c_str(), "B");
+  EXPECT_STREQ(ds->get_base_frame().get_name().c_str(), "B");
+  EXPECT_STREQ(ds->get_base_frame().get_reference_frame().c_str(), "B");
+  EXPECT_NEAR(ds->get_base_frame().get_pose().norm(), 1, tol);
 
   // setting the center should fail if it is incompatible with the base frame
   auto CinD = CartesianState::Random("C", "D");
-  EXPECT_THROW(ring.set_center(CinD), state_representation::exceptions::IncompatibleReferenceFramesException);
+  EXPECT_THROW(ds->set_parameter(make_shared_parameter("center", CartesianPose(CinD))),
+               state_representation::exceptions::IncompatibleReferenceFramesException);
 
   // updating the base frame should "move" the center frame but not change its magnitude
-  auto centerNorm = ring.get_center().get_pose().norm();
-  ring.set_base_frame(CinD);
+  auto centerNorm = ds->get_parameter_value<CartesianPose>("center").get_pose().norm();
+  ds->set_base_frame(CinD);
 
-  EXPECT_STREQ(ring.get_center().get_name().c_str(), "A");
-  EXPECT_STREQ(ring.get_center().get_reference_frame().c_str(), "C");
-  EXPECT_STREQ(ring.get_base_frame().get_name().c_str(), "C");
-  EXPECT_STREQ(ring.get_base_frame().get_reference_frame().c_str(), "D");
-  EXPECT_NEAR(ring.get_center().get_pose().norm(), centerNorm, tol);
+  EXPECT_STREQ(ds->get_parameter_value<CartesianPose>("center").get_name().c_str(), "A");
+  EXPECT_STREQ(ds->get_parameter_value<CartesianPose>("center").get_reference_frame().c_str(), "C");
+  EXPECT_STREQ(ds->get_base_frame().get_name().c_str(), "C");
+  EXPECT_STREQ(ds->get_base_frame().get_reference_frame().c_str(), "D");
+  EXPECT_NEAR(ds->get_parameter_value<CartesianPose>("center").get_pose().norm(), centerNorm, tol);
 
   // setting the center should succeed if it is expressed relative to the base frame
   auto BinC = CartesianState::Random("B", "C");
-  EXPECT_NO_THROW(ring.set_center(BinC));
-  EXPECT_STREQ(ring.get_center().get_name().c_str(), "B");
-  EXPECT_STREQ(ring.get_center().get_reference_frame().c_str(), "C");
+  EXPECT_NO_THROW(ds->set_parameter(make_shared_parameter("center", CartesianPose(BinC))));
+  EXPECT_STREQ(ds->get_parameter_value<CartesianPose>("center").get_name().c_str(), "B");
+  EXPECT_STREQ(ds->get_parameter_value<CartesianPose>("center").get_reference_frame().c_str(), "C");
 
   // setting the center should also succeed if it shares the same reference frame as the base frame
   auto BinD = CartesianState::Random("B", "D");
-  EXPECT_NO_THROW(ring.set_center(BinD));
-  EXPECT_STREQ(ring.get_center().get_name().c_str(), "B");
+  EXPECT_NO_THROW(ds->set_parameter(make_shared_parameter("center", CartesianPose(BinD))));
+  EXPECT_STREQ(ds->get_parameter_value<CartesianPose>("center").get_name().c_str(), "B");
   // the reference frame is still C, not D, because the center is internally represented relative to the base frame (C)
-  EXPECT_STREQ(ring.get_center().get_reference_frame().c_str(), "C");
+  EXPECT_STREQ(ds->get_parameter_value<CartesianPose>("center").get_reference_frame().c_str(), "C");
 }
 
 TEST_F(RingDSTest, SettersAndGetters) {
-  dynamical_systems::Ring ring(center);
+  ds->set_parameter(make_shared_parameter<CartesianPose>("center", center));
 
   auto pose = CartesianState::Random("C");
-  ring.set_center(pose);
-  auto pose2 = ring.get_center();
+  ds->set_parameter(make_shared_parameter("center", CartesianPose(pose)));
+  auto pose2 = ds->get_parameter_value<CartesianPose>("center");
   EXPECT_STREQ(pose.get_name().c_str(), pose2.get_name().c_str());
   EXPECT_STREQ(pose.get_reference_frame().c_str(), pose2.get_reference_frame().c_str());
   EXPECT_NEAR(pose.get_pose().norm(), pose2.get_pose().norm(), tol);
 
   // all other setters should store the value
-  ring.set_rotation_offset(Eigen::Quaterniond(1, 2, 3, 4).normalized());
-  EXPECT_NEAR(ring.get_rotation_offset().angularDistance(Eigen::Quaterniond(1, 2, 3, 4).normalized()), 0, tol);
+  ds->set_parameter(
+      make_shared_parameter("rotation_offset", CartesianPose("offset", Eigen::Quaterniond(1, 2, 3, 4).normalized())));
+  EXPECT_NEAR(ds->get_parameter_value<CartesianPose>("rotation_offset").get_orientation().angularDistance(
+      Eigen::Quaterniond(1, 2, 3, 4).normalized()), 0, tol);
 
-  ring.set_radius(1);
-  EXPECT_NEAR(ring.get_radius(), 1, tol);
+  ds->set_parameter(make_shared_parameter("radius", 1.0));
+  EXPECT_NEAR(ds->get_parameter_value<double>("radius"), 1.0, tol);
 
-  ring.set_width(2);
-  EXPECT_NEAR(ring.get_width(), 2, tol);
+  ds->set_parameter(make_shared_parameter("width", 2.0));
+  EXPECT_NEAR(ds->get_parameter_value<double>("width"), 2.0, tol);
 
-  ring.set_speed(3);
-  EXPECT_NEAR(ring.get_speed(), 3, tol);
+  ds->set_parameter(make_shared_parameter("speed", 3.0));
+  EXPECT_NEAR(ds->get_parameter_value<double>("speed"), 3.0, tol);
 
-  ring.set_field_strength(4);
-  EXPECT_NEAR(ring.get_field_strength(), 4, tol);
+  ds->set_parameter(make_shared_parameter("field_strength", 4.0));
+  EXPECT_NEAR(ds->get_parameter_value<double>("field_strength"), 4.0, tol);
 
-  ring.set_normal_gain(5);
-  EXPECT_NEAR(ring.get_normal_gain(), 5, tol);
+  ds->set_parameter(make_shared_parameter("normal_gain", 5.0));
+  EXPECT_NEAR(ds->get_parameter_value<double>("normal_gain"), 5.0, tol);
 
-  ring.set_angular_gain(6);
-  EXPECT_NEAR(ring.get_angular_gain(), 6, tol);
+  ds->set_parameter(make_shared_parameter("angular_gain", 6.0));
+  EXPECT_NEAR(ds->get_parameter_value<double>("angular_gain"), 6.0, tol);
 }