+++ CHANGELOG.md
@@ -13,7 +13,11 @@ Release Versions:
 - Add static method to create Parameter pointer (#226)
 - Templated get_value method for Parameter (#228)
 - Create DS interface and DS factory classes (#227)
+- Refactor Linear DS to PointAttractor DS with factory (#229)
+- Refactor Circular DS with factory (#230)
+- Refactor Ring DS with factory and remove old DS base class (#231)
 - Add support for copy module in Python bindings (#232)
+- Refactor dynamical systems using factory pattern (#233) 
 
 ### Pending TODOs for the next release